chore: extract download logic to helper methods
We are discarding the stale parameter, because the stale version is set
when the installer is required, so we simply download what is requested
