+++ e2e/_suites/fleet/services_test.go
@@ -7,24 +7,23 @@ package main
 import (
 	"os"
 	"path"
+	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
 
-var testVersion = agentVersionBase
-
 func TestDownloadAgentBinary(t *testing.T) {
 	artifact := "elastic-agent"
 	beatsDir := path.Join("..", "..", "_testresources", "beats")
-	distributionsDir := path.Join(beatsDir, "x-pack", "elastic-agent", "build", "distributions")
+	distributionsDir, _ := filepath.Abs(path.Join(beatsDir, "x-pack", "elastic-agent", "build", "distributions"))
 	version := "8.0.0-SNAPSHOT"
 
 	t.Run("Fetching non-existent binary from local Beats dir throws an error", func(t *testing.T) {
 		defer os.Unsetenv("BEATS_LOCAL_PATH")
 		os.Setenv("BEATS_LOCAL_PATH", beatsDir)
 
-		_, err := downloadAgentBinary("foo_fileName", artifact, version, false)
+		_, err := downloadAgentBinary("foo_fileName", artifact, version)
 		assert.NotNil(t, err)
 	})
 
@@ -33,10 +32,11 @@ func TestDownloadAgentBinary(t *testing.T) {
 		os.Setenv("BEATS_LOCAL_PATH", beatsDir)
 
 		artifactName := "elastic-agent-8.0.0-SNAPSHOT-x86_64.rpm"
+		expectedFilePath := path.Join(distributionsDir, artifactName)
 
-		downloadedFilePath, err := downloadAgentBinary(artifactName, artifact, version, false)
+		downloadedFilePath, err := downloadAgentBinary(artifactName, artifact, version)
 		assert.Nil(t, err)
-		assert.Equal(t, downloadedFilePath, path.Join(distributionsDir, artifactName))
+		assert.Equal(t, downloadedFilePath, expectedFilePath)
 	})
 
 	t.Run("Fetching DEB binary from local Beats dir", func(t *testing.T) {
@@ -44,10 +44,11 @@ func TestDownloadAgentBinary(t *testing.T) {
 		os.Setenv("BEATS_LOCAL_PATH", beatsDir)
 
 		artifactName := "elastic-agent-8.0.0-SNAPSHOT-amd64.deb"
+		expectedFilePath := path.Join(distributionsDir, artifactName)
 
-		downloadedFilePath, err := downloadAgentBinary(artifactName, artifact, version, false)
+		downloadedFilePath, err := downloadAgentBinary(artifactName, artifact, version)
 		assert.Nil(t, err)
-		assert.Equal(t, downloadedFilePath, path.Join(distributionsDir, artifactName))
+		assert.Equal(t, downloadedFilePath, expectedFilePath)
 	})
 
 	t.Run("Fetching TAR binary from local Beats dir", func(t *testing.T) {
@@ -55,10 +56,11 @@ func TestDownloadAgentBinary(t *testing.T) {
 		os.Setenv("BEATS_LOCAL_PATH", beatsDir)
 
 		artifactName := "elastic-agent-8.0.0-SNAPSHOT-linux-amd64.tar.gz"
+		expectedFilePath := path.Join(distributionsDir, artifactName)
 
-		downloadedFilePath, err := downloadAgentBinary(artifactName, artifact, version, false)
+		downloadedFilePath, err := downloadAgentBinary(artifactName, artifact, version)
 		assert.Nil(t, err)
-		assert.Equal(t, downloadedFilePath, path.Join(distributionsDir, artifactName))
+		assert.Equal(t, downloadedFilePath, expectedFilePath)
 	})
 
 	t.Run("Fetching Docker binary from local Beats dir", func(t *testing.T) {
@@ -66,10 +68,11 @@ func TestDownloadAgentBinary(t *testing.T) {
 		os.Setenv("BEATS_LOCAL_PATH", beatsDir)
 
 		artifactName := "elastic-agent-8.0.0-SNAPSHOT-linux-amd64.docker.tar.gz"
+		expectedFilePath := path.Join(distributionsDir, artifactName)
 
-		downloadedFilePath, err := downloadAgentBinary(artifactName, artifact, version, false)
+		downloadedFilePath, err := downloadAgentBinary(artifactName, artifact, version)
 		assert.Nil(t, err)
-		assert.Equal(t, downloadedFilePath, path.Join(distributionsDir, artifactName))
+		assert.Equal(t, downloadedFilePath, expectedFilePath)
 	})
 
 	t.Run("Fetching ubi8 Docker binary from local Beats dir", func(t *testing.T) {
@@ -77,261 +80,10 @@ func TestDownloadAgentBinary(t *testing.T) {
 		os.Setenv("BEATS_LOCAL_PATH", beatsDir)
 
 		artifactName := "elastic-agent-ubi8-8.0.0-SNAPSHOT-linux-amd64.docker.tar.gz"
+		expectedFilePath := path.Join(distributionsDir, artifactName)
 
-		downloadedFilePath, err := downloadAgentBinary(artifactName, artifact, version, false)
+		downloadedFilePath, err := downloadAgentBinary(artifactName, artifact, version)
 		assert.Nil(t, err)
-		assert.Equal(t, downloadedFilePath, path.Join(distributionsDir, artifactName))
-	})
-}
-
-func TestGetGCPBucketCoordinates_Commits(t *testing.T) {
-	artifact := "elastic-agent"
-	version := testVersion
-
-	t.Run("Fetching commits bucket for RPM package", func(t *testing.T) {
-		defer os.Unsetenv("GITHUB_CHECK_SHA1")
-		os.Setenv("GITHUB_CHECK_SHA1", "0123456789")
-
-		fileName := "elastic-agent-" + testVersion + "-x86_64.rpm"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, false)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "commits/0123456789")
-		assert.Equal(t, object, "elastic-agent/elastic-agent-"+testVersion+"-x86_64.rpm")
-	})
-	t.Run("Fetching commits bucket for stale RPM package in the snapshots folder", func(t *testing.T) {
-		defer os.Unsetenv("GITHUB_CHECK_SHA1")
-		os.Setenv("GITHUB_CHECK_SHA1", "0123456789")
-
-		fileName := "elastic-agent-" + testVersion + "-x86_64.rpm"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, true)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "snapshots/elastic-agent")
-		assert.Equal(t, object, "elastic-agent-"+testVersion+"-x86_64.rpm")
-	})
-
-	t.Run("Fetching commits bucket for DEB package", func(t *testing.T) {
-		defer os.Unsetenv("GITHUB_CHECK_SHA1")
-		os.Setenv("GITHUB_CHECK_SHA1", "0123456789")
-
-		fileName := "elastic-agent-" + testVersion + "-amd64.deb"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, false)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "commits/0123456789")
-		assert.Equal(t, object, "elastic-agent/elastic-agent-"+testVersion+"-amd64.deb")
-	})
-	t.Run("Fetching commits bucket for stale DEB package in the snapshots folder", func(t *testing.T) {
-		defer os.Unsetenv("GITHUB_CHECK_SHA1")
-		os.Setenv("GITHUB_CHECK_SHA1", "0123456789")
-
-		fileName := "elastic-agent-" + testVersion + "-amd64.deb"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, true)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "snapshots/elastic-agent")
-		assert.Equal(t, object, "elastic-agent-"+testVersion+"-amd64.deb")
-	})
-
-	t.Run("Fetching commits bucket for TAR package adds OS to fileName and object", func(t *testing.T) {
-		defer os.Unsetenv("GITHUB_CHECK_SHA1")
-		os.Setenv("GITHUB_CHECK_SHA1", "0123456789")
-
-		fileName := "elastic-agent-" + testVersion + "-linux-x86_64.tar.gz"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, false)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "commits/0123456789")
-		assert.Equal(t, object, "elastic-agent/elastic-agent-"+testVersion+"-linux-x86_64.tar.gz")
-	})
-	t.Run("Fetching commits bucket for stale TAR package adds OS to fileName and object in the snapshots folder", func(t *testing.T) {
-		defer os.Unsetenv("GITHUB_CHECK_SHA1")
-		os.Setenv("GITHUB_CHECK_SHA1", "0123456789")
-
-		fileName := "elastic-agent-" + testVersion + "-linux-x86_64.tar.gz"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, true)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "snapshots/elastic-agent")
-		assert.Equal(t, object, "elastic-agent-"+testVersion+"-linux-x86_64.tar.gz")
-	})
-}
-
-func TestGetGCPBucketCoordinates_CommitsForAPullRequest(t *testing.T) {
-	artifact := "elastic-agent"
-	version := "pr-23456"
-
-	t.Run("Fetching commits bucket for RPM package", func(t *testing.T) {
-		defer os.Unsetenv("GITHUB_CHECK_SHA1")
-		os.Setenv("GITHUB_CHECK_SHA1", "0123456789")
-
-		fileName := "elastic-agent-" + testVersion + "-x86_64.rpm"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, false)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "pull-requests/pr-23456")
-		assert.Equal(t, object, "elastic-agent/elastic-agent-"+testVersion+"-x86_64.rpm")
-	})
-	t.Run("Fetching commits bucket for stale RPM package in the snapshots folder", func(t *testing.T) {
-		defer os.Unsetenv("GITHUB_CHECK_SHA1")
-		os.Setenv("GITHUB_CHECK_SHA1", "0123456789")
-
-		fileName := "elastic-agent-" + testVersion + "-x86_64.rpm"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, true)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "snapshots/elastic-agent")
-		assert.Equal(t, object, "elastic-agent-"+testVersion+"-x86_64.rpm")
-	})
-
-	t.Run("Fetching commits bucket for DEB package", func(t *testing.T) {
-		defer os.Unsetenv("GITHUB_CHECK_SHA1")
-		os.Setenv("GITHUB_CHECK_SHA1", "0123456789")
-
-		fileName := "elastic-agent-" + testVersion + "-amd64.deb"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, false)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "pull-requests/pr-23456")
-		assert.Equal(t, object, "elastic-agent/elastic-agent-"+testVersion+"-amd64.deb")
-	})
-	t.Run("Fetching commits bucket for stale DEB package in the snapshots folder", func(t *testing.T) {
-		defer os.Unsetenv("GITHUB_CHECK_SHA1")
-		os.Setenv("GITHUB_CHECK_SHA1", "0123456789")
-
-		fileName := "elastic-agent-" + testVersion + "-amd64.deb"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, true)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "snapshots/elastic-agent")
-		assert.Equal(t, object, "elastic-agent-"+testVersion+"-amd64.deb")
-	})
-
-	t.Run("Fetching commits bucket for TAR package adds OS to fileName and object", func(t *testing.T) {
-		defer os.Unsetenv("GITHUB_CHECK_SHA1")
-		os.Setenv("GITHUB_CHECK_SHA1", "0123456789")
-
-		fileName := "elastic-agent-" + testVersion + "-linux-x86_64.tar.gz"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, false)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "pull-requests/pr-23456")
-		assert.Equal(t, object, "elastic-agent/elastic-agent-"+testVersion+"-linux-x86_64.tar.gz")
-	})
-	t.Run("Fetching commits bucket for stale TAR package adds OS to fileName and object in the snapshots folder", func(t *testing.T) {
-		defer os.Unsetenv("GITHUB_CHECK_SHA1")
-		os.Setenv("GITHUB_CHECK_SHA1", "0123456789")
-
-		fileName := "elastic-agent-" + testVersion + "-linux-x86_64.tar.gz"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, true)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "snapshots/elastic-agent")
-		assert.Equal(t, object, "elastic-agent-"+testVersion+"-linux-x86_64.tar.gz")
-	})
-}
-
-func TestGetGCPBucketCoordinates_PullRequests(t *testing.T) {
-	artifact := "elastic-agent"
-	version := "pr-23456"
-
-	t.Run("Fetching commits bucket for RPM package", func(t *testing.T) {
-		fileName := "elastic-agent-" + testVersion + "-x86_64.rpm"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, false)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "pull-requests/pr-23456")
-		assert.Equal(t, object, "elastic-agent/elastic-agent-"+testVersion+"-x86_64.rpm")
-	})
-	t.Run("Fetching commits bucket for stale RPM package in the snapshots folder", func(t *testing.T) {
-		fileName := "elastic-agent-" + testVersion + "-x86_64.rpm"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, true)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "snapshots/elastic-agent")
-		assert.Equal(t, object, "elastic-agent-"+testVersion+"-x86_64.rpm")
-	})
-
-	t.Run("Fetching commits bucket for DEB package", func(t *testing.T) {
-		fileName := "elastic-agent-" + testVersion + "-amd64.deb"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, false)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "pull-requests/pr-23456")
-		assert.Equal(t, object, "elastic-agent/elastic-agent-"+testVersion+"-amd64.deb")
-	})
-	t.Run("Fetching commits bucket for stale DEB package in the snapshots folder", func(t *testing.T) {
-		fileName := "elastic-agent-" + testVersion + "-amd64.deb"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, true)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "snapshots/elastic-agent")
-		assert.Equal(t, object, "elastic-agent-"+testVersion+"-amd64.deb")
-	})
-
-	t.Run("Fetching commits bucket for TAR package adds OS to fileName and object", func(t *testing.T) {
-		fileName := "elastic-agent-" + testVersion + "-linux-x86_64.tar.gz"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, false)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "pull-requests/pr-23456")
-		assert.Equal(t, object, "elastic-agent/elastic-agent-"+testVersion+"-linux-x86_64.tar.gz")
-	})
-	t.Run("Fetching commits bucket for stale TAR package adds OS to fileName and object in the snapshots folder", func(t *testing.T) {
-		fileName := "elastic-agent-" + testVersion + "-linux-x86_64.tar.gz"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, true)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "snapshots/elastic-agent")
-		assert.Equal(t, object, "elastic-agent-"+testVersion+"-linux-x86_64.tar.gz")
-	})
-}
-
-func TestGetGCPBucketCoordinates_Snapshots(t *testing.T) {
-	artifact := "elastic-agent"
-	version := testVersion
-
-	t.Run("Fetching commits bucket for RPM package", func(t *testing.T) {
-		fileName := "elastic-agent-" + testVersion + "-x86_64.rpm"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, false)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "snapshots/elastic-agent")
-		assert.Equal(t, object, "elastic-agent-"+testVersion+"-x86_64.rpm")
-	})
-
-	t.Run("Fetching commits bucket for DEB package", func(t *testing.T) {
-		fileName := "elastic-agent-" + testVersion + "-amd64.deb"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, false)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "snapshots/elastic-agent")
-		assert.Equal(t, object, "elastic-agent-"+testVersion+"-amd64.deb")
-	})
-	t.Run("Fetching commits bucket for stale DEB package", func(t *testing.T) {
-		fileName := "elastic-agent-" + testVersion + "-amd64.deb"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, true)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "snapshots/elastic-agent")
-		assert.Equal(t, object, "elastic-agent-"+testVersion+"-amd64.deb")
-	})
-
-	t.Run("Fetching commits bucket for TAR package adds OS to fileName and object", func(t *testing.T) {
-		fileName := "elastic-agent-" + testVersion + "-linux-x86_64.tar.gz"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, false)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "snapshots/elastic-agent")
-		assert.Equal(t, object, "elastic-agent-"+testVersion+"-linux-x86_64.tar.gz")
-	})
-	t.Run("Fetching commits bucket for stale TAR package adds OS to fileName and object", func(t *testing.T) {
-		fileName := "elastic-agent-" + testVersion + "-linux-x86_64.tar.gz"
-
-		bucket, prefix, object := getGCPBucketCoordinates(fileName, artifact, version, true)
-		assert.Equal(t, bucket, "beats-ci-artifacts")
-		assert.Equal(t, prefix, "snapshots/elastic-agent")
-		assert.Equal(t, object, "elastic-agent-"+testVersion+"-linux-x86_64.tar.gz")
+		assert.Equal(t, downloadedFilePath, expectedFilePath)
 	})
 }
